backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.7.8
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
sqlmodel==0.0.16
SQLAlchemy==2.0.32
psycopg[binary]==3.2.10
python-dotenv==1.0.0
supabase==2.6.0
PyJWT[crypto]==2.9.0
cryptography==43.0.1